set policies

+++ app/Helpers/DB/Controller/WithTranslation/DeleteTrait.php
@@ -5,10 +5,10 @@ namespace App\Helpers\DB\Controller\WithTranslation;
 use App\Support\ApiResponse;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
-trait DeleteTrait 
+trait DeleteTrait
 {
     public function destroy($id):JsonResponse {
-
+        $this->authorize('delete', $this->model);
         $result = $this->service->destroy($id);
 
         return ApiResponse::success(
